Move LSPSettings an ClientConfig to a separate module
- start modularization to split the plugin into manageable pieces
- prevent mypy from complaining about dynamically added attributes
